+++ CHANGES.md
@@ -18,6 +18,7 @@
 * moved checkboxes before titles in Settings
 * updated wiki screenshots
 * fixed to call `load_options()` in `Base::init()` to properly utilize options
+* add red (#f00) warning dashicon in Settings for repo with malformed header URI
 
 #### 6.2.2 / 2017-02-09
 * fixed for updating via webhook from GitHub tagged release, declare branch as `master`